Extract GiveGiftRibbonToParty from file about size records
