+++ include/give_gift_ribbon_to_party.h
@@ -1,0 +1,6 @@
+#ifndef GUARD_GIVE_GIFT_RIBBON_TO_PARTY_H
+#define GUARD_GIVE_GIFT_RIBBON_TO_PARTY_H
+
+void GiveGiftRibbonToParty(u8 index, u8 ribbonId);
+
+#endif // GUARD_GIVE_GIFT_RIBBON_TO_PARTY_H
