+++ include/pokemon_size_record.h
@@ -9,6 +9,4 @@ void InitLotadSizeRecord(void);
 void GetLotadSizeRecordInfo(void);
 void CompareLotadSize(void);
 
-void GiveGiftRibbonToParty(u8 index, u8 ribbonId);
-
 #endif // GUARD_POKEMON_SIZE_RECORD_H
